requirements.txt
dbt-postgres>=1.8.0
pandas>=2.0.0
pyarrow>=14.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
tqdm>=4.66.0
